Improve link-breaking strategy in export/import sheets code
Exporting/importing a sheet can cause more than one Excel link to be
added to the destination workbook, if some of the formulas on the sheet
in question contain links to a different workbook.  So, store the list
of linked workbooks in the destination workbook before and after copying
sheets, and attempt to break any new links that were introduced during
the copy.

Also, detect and gracefully handle cases where links failed to be
broken.